vehiclemanagerbackend/requirements.txt
aniso8601==7.0.0
asgiref==3.5.2
Django==3.2.15
django-cors-headers==3.13.0
django-graphql-jwt==0.3.4
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
PyJWT==2.4.0
pytz==2022.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
